backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
python-multipart==0.0.6

# AI & LLM
anthropic==0.18.0

# Voice Processing
edge-tts==6.1.9
openai-whisper==20231117
pydub==0.25.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
motor==3.3.2
redis==5.0.1

# Background Tasks
celery==5.3.6

# Real-time Collaboration
ypy-websocket==0.12.1
y-py==0.6.2

# Data Validation & Serialization
pydantic==2.6.0
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
